+++ test/test_parser.cpp
@@ -1750,8 +1750,7 @@ TEST(Parser_Backlinks)
     CHECK_EQUAL(message, "Comparing a list property to 'null' is not supported");
     CHECK_THROW_ANY_GET_MESSAGE(verify_query(test_context, items, "@links.class_Person.fav_item == NULL", 1), message);
     CHECK_EQUAL(message, "Comparing a list property to 'null' is not supported");
-    CHECK_THROW_ANY_GET_MESSAGE(verify_query(test_context, items, "@links.@count > 0", 1), message);
-    CHECK_EQUAL(message, "'@links' must be proceeded by type name and a property name");
+    CHECK_THROW_ANY(verify_query(test_context, items, "@links.attr. > 0", 1));
     CHECK_THROW_ANY_GET_MESSAGE(verify_query(test_context, items, "@links.class_Factory.items > 0", 1), message);
     CHECK_EQUAL(message, "No property 'items' found in type 'Factory' which links to type 'Items'");
     CHECK_THROW_ANY_GET_MESSAGE(verify_query(test_context, items, "@links.class_Person.artifacts > 0", 1), message);
@@ -1795,6 +1794,121 @@ TEST(Parser_Backlinks)
     p = realm::parser::parse("purchasers.@max.money >= 20").predicate;
     realm::query_builder::apply_predicate(q, p, args, mapping_with_prefix);
     CHECK_EQUAL(q.count(), 3);
+}
+
+
+TEST(Parser_BacklinkCount)
+{
+    Group g;
+
+    TableRef items = g.add_table("class_Items");
+    size_t item_id_col = items->add_column(type_Int, "item_id");
+    size_t item_link_col = items->add_column_link(type_Link, "self", *items);
+    size_t double_col = items->add_column(type_Double, "double_col");
+    items->add_empty_row(5);
+    for (size_t i = 0; i < items->size(); ++i) {
+        items->set_int(item_id_col, i, i);
+        items->set_link(item_link_col, i, i);
+        items->set_double(double_col, i, double(i) + 0.5);
+    }
+    items->nullify_link(item_link_col, 4); // last item will have a total of 0 backlinks
+
+    TableRef t = g.add_table("class_Person");
+    size_t id_col = t->add_column(type_Int, "customer_id");
+    size_t items_col = t->add_column_link(type_LinkList, "items", *items);
+    size_t fav_col = t->add_column_link(type_Link, "fav_item", *items);
+    size_t float_col = t->add_column(type_Float, "float_col");
+    t->add_empty_row(3);
+    for (size_t i = 0; i < t->size(); ++i) {
+        t->set_int(id_col, i, i);
+        t->set_link(fav_col, i, i);
+        t->set_float(float_col, i, float(i) + 0.5f);
+    }
+
+    LinkViewRef list_0 = t->get_linklist(items_col, 0);
+    list_0->add(0);
+    list_0->add(1);
+    list_0->add(2);
+
+    LinkViewRef list_1 = t->get_linklist(items_col, 1);
+    for (size_t i = 0; i < 10; ++i) {
+        list_1->add(0);
+    }
+
+    LinkViewRef list_2 = t->get_linklist(items_col, 2);
+    list_2->add(2);
+    list_2->add(2);
+
+    verify_query(test_context, items, "@links.@count == 0", 1);
+    verify_query(test_context, items, "@links.@count == 0 && item_id == 4", 1);
+    verify_query(test_context, items, "@links.@count == 1", 1);
+    verify_query(test_context, items, "@links.@count == 1 && item_id == 3", 1);
+    verify_query(test_context, items, "@links.@count == 5", 1);
+    verify_query(test_context, items, "@links.@count == 5 && item_id == 2", 1);
+    verify_query(test_context, items, "@links.@count == 3", 1);
+    verify_query(test_context, items, "@links.@count == 3 && item_id == 1", 1);
+    verify_query(test_context, items, "@links.@count == 13", 1);
+    verify_query(test_context, items, "@links.@count == 13 && item_id == 0", 1);
+
+    // @size is still a synonym to @count
+    verify_query(test_context, items, "@links.@size == 0", 1);
+    verify_query(test_context, items, "@links.@size == 0 && item_id == 4", 1);
+
+    // backlink count through forward links
+    verify_query(test_context, t, "fav_item.@links.@count == 3 && fav_item.item_id == 1", 1);
+    verify_query(test_context, t, "fav_item.@links.@count == 13 && fav_item.item_id == 0", 1);
+
+    // backlink count through backlinks first
+    verify_query(test_context, items, "@links.class_Items.self.@links.@count == 1 && item_id == 3", 1);
+    verify_query(test_context, items, "@links.class_Person.items.@links.@count == 0", 5);
+
+    // backlink count through backlinks and forward links
+    verify_query(test_context, items, "@links.class_Person.fav_item.items.@links.@count == 1 && item_id == 3", 1);
+    verify_query(test_context, items, "@links.class_Person.fav_item.fav_item.@links.@count == 1 && item_id == 3", 1);
+
+    // backlink count compared to int
+    verify_query(test_context, items, "@links.@count == 0", 1);
+    verify_query(test_context, items, "@links.@count >= item_id", 3); // 3 items have an id less than their backlink count
+    verify_query(test_context, items, "@links.@count >= @links.class_Person.fav_item.customer_id", 3);
+
+    // backlink count compared to double
+    verify_query(test_context, items, "@links.@count == 0.0", 1);
+    verify_query(test_context, items, "@links.@count >= double_col", 3);
+
+    // backlink count compared to float
+    verify_query(test_context, items, "@links.@count >= @links.class_Person.fav_item.float_col", 3);
+
+    // backlink count compared to link count
+    verify_query(test_context, items, "@links.@count >= self.@count", 5);
+    verify_query(test_context, t, "items.@count >= items.@links.@count", 1);
+    verify_query(test_context, t, "items.@count >= fav_item.@links.@count", 1); // equivilent to above
+
+    // all backlinks count compared to single column backlink count
+    // this is essentially checking if a single column contains all backlinks of a object
+    verify_query(test_context, items, "@links.@count == @links.class_Person.fav_item.@count", 1); // item 5 (0 links)
+    verify_query(test_context, items, "@links.@count == @links.class_Person.items.@count", 1); // item 5 (0 links)
+    verify_query(test_context, items, "@links.@count == @links.class_Items.self.@count", 2); // items 4,5 (1,0 links)
+
+    std::string message;
+    // backlink count requires comparison to a numeric type
+    CHECK_THROW_ANY_GET_MESSAGE(verify_query(test_context, items, "@links.@count == 'string'", -1), message);
+    CHECK_EQUAL(message, "Cannot convert string 'string' to type 'Int'");
+    CHECK_THROW_ANY_GET_MESSAGE(verify_query(test_context, items, "@links.@count == 2018-04-09@14:21:0", -1), message);
+    CHECK_EQUAL(message, "Attempting to compare a numeric property to a non-numeric value");
+
+    // no suffix after @links.@count is allowed
+    CHECK_THROW_ANY(verify_query(test_context, items, "@links.@count.item_id == 0", -1));
+    CHECK_THROW_ANY(verify_query(test_context, items, "@links.@count.@avg.items_id == 0", -1));
+
+    // other aggregate operators are not supported
+    CHECK_THROW_ANY(verify_query(test_context, items, "@links.@avg == 1", -1));
+    CHECK_THROW_ANY(verify_query(test_context, items, "@links.@sum == 1", -1));
+    CHECK_THROW_ANY(verify_query(test_context, items, "@links.@min == 1", -1));
+    CHECK_THROW_ANY(verify_query(test_context, items, "@links.@max == 1", -1));
+    CHECK_THROW_ANY(verify_query(test_context, items, "@links.@avg.item_id == 1", -1));
+    CHECK_THROW_ANY(verify_query(test_context, items, "@links.@sum.item_id == 1", -1));
+    CHECK_THROW_ANY(verify_query(test_context, items, "@links.@min.item_id == 1", -1));
+    CHECK_THROW_ANY(verify_query(test_context, items, "@links.@max.item_id == 1", -1));
 }