+++ CHANGELOG.md
@@ -14,6 +14,8 @@
     - Allow an arbitrary prefix on backlink class names of @links queries.
       This will allow users to query unnamed backlinks using the `@links.Class.property` syntax.
     - Case insensitive `nil` is now recognised as a synonym to `NULL`.
+    - Add support for `@links.@count` which gives the count of all backlinks to an object.
+      See Issue [#3003](https://github.com/realm/realm-core/issues/3003).
 
 -----------